be/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pymongo==4.8.0
python-dotenv==1.0.1
